3 major milestones: 1. agents can start with completely random seeds 2. fitness is now evaluated solely on captures 3. both 3-4 task problems have been verified to work

+++ src/com/synthverse/synthscape/core/ExperimentReporter.java
@@ -158,7 +158,7 @@ public class ExperimentReporter implements Constants {
 
     }
 
-    public void reportEvent(int simulationNumber, int generation, Species species, int agentId, int step, int x, int y,
+    public void reportEvent(long simulationNumber, int generation, Species species, int agentId, int step, int x, int y,
 	    Event event, String source, String destination) {
 	try {
 	    if (simulation.isReportEvents()) {
